requirements.txt
matplotlib==3.7.4
numpy==1.24.3
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==10.0.1
torch==1.10.0
torchvision==0.11.0
tqdm==4.65.0
